+++ libs/rxjs/array/src/lib/find-all.ts
@@ -6,10 +6,11 @@
 import { OperatorFunction } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { PredicateFn } from '../types/generic-methods';
-import { ArrayOrSet } from '../types/array-set';
 
 /**
- * Returns an Observable array of truthy values from a source array
+ * Returns an Observable array of truthy values from a source `Array` or `Set`
+ * When working with data, if the array contains numbers `0` will be returned as a value to the [[PredicateFn]], but all
+ * other falsy values will be ignored
  *
  * @category Filter
  *
@@ -33,7 +34,7 @@ import { ArrayOrSet } from '../types/array-set';
  *
  * @returns An Observable that emits an array containing all truthy values from a source array
  */
-export function findAll<T extends unknown>(predicate?: PredicateFn<T>): OperatorFunction<ArrayOrSet<T>, T[]> {
+export function findAll<T extends unknown>(predicate?: PredicateFn<T>): OperatorFunction<Iterable<T>, T[]> {
   return (source) =>
     source.pipe(
       map((value) =>