+++ libs/rxjs/array/src/lib/from-set.ts
@@ -2,10 +2,9 @@
  * @packageDocumentation
  * @module Array
  */
-import { isObservable, Observable, ObservableInput, throwError } from 'rxjs';
-import { catchError, finalize, map, takeWhile, tap } from 'rxjs/operators';
-import { fromPromise, isPromise } from 'rxjs/internal-compatibility';
-import { flatMapSet } from '../utils/array-set';
+import { Observable, Subscribable } from 'rxjs';
+import { finalize, map, takeWhile, tap } from 'rxjs/operators';
+import { createOrReturnObservable, flatMapSet } from '../utils/internal';
 
 /**
  * Returns an Observable that emits an `Array` from a `Set`
@@ -14,7 +13,7 @@ import { flatMapSet } from '../utils/array-set';
  *
  * @typeParam T The type of value contained in the `Set`
  *
- * @param args Input to create the emit values from, can be argument list of `Set`, an array of `Set` or an Observable
+ * @param input Input to create the emit values from, can be argument list of `Set`, an array of `Set` or an Observable
  *   or Promise source
  *
  * @example
@@ -28,45 +27,22 @@ import { flatMapSet } from '../utils/array-set';
  * @returns Observable that emits an `Array` from the input `Set`
  */
 export function fromSet<T extends unknown>(
-  ...args: (ObservableInput<Set<T>[] | Set<T>> | Set<T>[] | Set<T>)[]
+  input: Subscribable<Iterable<Set<T>> | Set<T>> | Iterable<Set<T>> | Set<T>,
 ): Observable<T[]> {
   return new Observable<T[]>((subscriber) => {
-    if (isObservable(args[0])) {
-      (args[0] as Observable<Set<T>[] | Set<T>>)
-        .pipe(
-          takeWhile(() => !subscriber.closed),
-          map<Set<T>[] | Set<T>, T[][]>((value) => flatMapSet(value)),
-          tap((value) => {
-            for (let i = 0; i < value.length; i++) {
-              subscriber.next(value[i]);
-            }
-            !subscriber.closed && subscriber.complete();
-          }),
-        )
-        .subscribe();
-    } else if (isPromise(args[0])) {
-      fromPromise(args[0] as Promise<Set<T>[] | Set<T>>)
-        .pipe(
-          map<Set<T>[] | Set<T>, T[][]>((value) => flatMapSet(value)),
-          tap((value) => {
-            for (let i = 0; i < value.length; i++) {
-              subscriber.next(value[i]);
-            }
-          }),
-          catchError((error) => {
-            subscriber.error(error);
-            return throwError(error);
-          }),
-          finalize(() => !subscriber.closed && subscriber.complete()),
-        )
-        .subscribe();
-    } else {
-      const value = Array.isArray(args[0]) ? flatMapSet(args[0]) : flatMapSet([...(args as Set<T>[])]);
-      for (let i = 0; i < value.length; i++) {
-        subscriber.next(value[i]);
-      }
-      !subscriber.closed && subscriber.complete();
-    }
+    createOrReturnObservable(input)
+      .pipe(
+        takeWhile(() => !subscriber.closed),
+        map<Iterable<Set<T>> | Set<T>, T[][]>((value) => flatMapSet(value)),
+        tap((value) => {
+          for (let i = 0; i < value.length; i++) {
+            subscriber.next(value[i] as T[]);
+          }
+        }),
+        finalize(() => !subscriber.closed && subscriber.complete()),
+      )
+      .subscribe();
+
     /* istanbul ignore next-line */
     return () => !subscriber.closed && subscriber.complete();
   });