+++ libs/rxjs/array/src/lib/reverse.ts
@@ -5,7 +5,6 @@
 
 import { OperatorFunction } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { ArrayOrSet } from '../types/array-set';
 
 /**
  * Returns an Observable that emits array taking the source and running the result of Array.reverse
@@ -24,6 +23,6 @@ import { ArrayOrSet } from '../types/array-set';
  *
  * @returns Observable that emits an array which is reversed from the source array
  */
-export function reverse<T extends unknown>(): OperatorFunction<ArrayOrSet<T>, T[]> {
+export function reverse<T extends unknown>(): OperatorFunction<Iterable<T>, T[]> {
   return (source) => source.pipe(map((value) => [...value].reverse()));
 }